fix: duplication reduction
